+++ lib/AppInfo/Application.php
@@ -7,6 +7,7 @@ declare(strict_types=1);
 
 namespace OCA\DeckRecurrence\AppInfo;
 
+use OCA\DeckRecurrence\Notification\Notifier;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
@@ -20,6 +21,7 @@ class Application extends App implements IBootstrap {
 	}
 
 	public function register(IRegistrationContext $context): void {
+		$context->registerNotifierService(Notifier::class);
 	}
 
 	public function boot(IBootContext $context): void {